Image editor and comments

+++ resources/js/components/content/GalleryComponent.vue
@@ -1,47 +1,69 @@
+<!--Displays the contents of each gallery with the image set as background for better styling.
+    Each of those elements has click function wich display the content as a single full screan element.
+    When the logged in user is administrator the content editing tools are shown. -->
 <template>
     <div class="row flex-xl-nowrap">
         <div class="col-12 pl-0 pr-0">
+<!--The following <div> is shown only if "adminMode" parameter is set to true.
+    He holds <a> with a click event that open the form to insert new records to the DB.
+    On the click event the prevent method was called to prevent the default behavior of the <a>
+    The href attribute is empty to keep the pointer:cursor behavior of the <a> -->
             <div v-if="adminMode" class="col-12">
                 <a href="" @click.prevent="showModal()"><i class="fas fa-file-import fa-3x green"></i></a>
             </div>
-            <h2 class="text-center dwhite">Галерия: {{this.$route.params.title}}</h2>
-            <div>
-                <div v-for="(photo, index) in displayedGallery" :key="photo.photo">
-                    <div class="photoGallery float-left" @click="singlePhotoModal(index)" :style="{ 'background-image': 'url(/storage/galleries/' + photo.photos + ')' }">
-                        <a v-if="adminMode && index > 0" href="" @click.prevent="deletePhoto(photo.id)">
-                            <i class="fas fa-trash red float-right fa-2x"></i>&#160;</a>
-                    </div>
+
+            <h2 class="text-center dwhite">Галерия: {{this.$route.params.title}}</h2><!-- Display the router parameter title -->
+
+<!--The following <div> is multiplied by Vuejs v-for directive for each record form "displayedGallery" parameter.
+Size is set to "col-3" grid proportion with zero padding to show 4 records per row.  -->
+            <div class="col-3 d-inline-block pl-0 pr-0" v-for="(photo, index) in displayedGallery" :key="photo.photo">
+<!--The following <div> has bind inline style background-image with the image name got from the DB -->
+                <div class="photoGallery" @click="singlePhotoModal(index)" :style="{ 'background-image': 'url(/storage/images/' + photo.photos + ')' }">
+                </div>
+<!--The following <a> checks is the user an administrator and is the element the second record from the DB.
+    if it is the <a> is displayed on the document with click event that calls "deletePhoto" passing the ID of the record. -->
+                <a v-if="adminMode && index > 0" href="" @click.prevent="deletePhoto(photo.id)">
+                        <i class="fas fa-trash red fa-lg"></i>&#160;</a>
+<!--The following <a> checks is the user an administrator if it is, the <a> is displayed on the document.
+    The binding href attribute is a path with the name of the image file to the Vuejs component which can edit it. -->
+                <a :href="'/Снимки/' + photo.photos" v-if="adminMode"><i class="fas fa-cut blue fa-lg"></i>&#160;</a>
+            </div>
+<!--The following <div> is displayed only if the records from the DB are more than the integer provided in the "loadedPhoto" parameter. -->
+            <div class="col-12 d-inline-block mt-3 dwhite" v-if="allPhotos > loadedPhoto">
+<!--The following <div> center  the <span> elements. The "galleryMenu" class is added for easier styling -->
+                <div class="text-center galleryMenu">
+                    <span>Общо снимки: {{ allPhotos }}</span><!-- Display the number of all records. -->
+                    <span>Показани: {{ loadedPhoto }}</span><!-- Display the number of shown records -->
+                    <span @click="loadMorePhoto()">Покажи още</span><!-- Call loadMorePhoto method -->
                 </div>
             </div>
-
-            <div :class="'col-12 d-inline-block mt-3 dwhite '" v-if="allPhotos > loadedPhoto">
-                <p class="text-center galleryMenu">
-                    <span>Общо снимки: {{ allPhotos }}</span><span>Показани: {{ loadedPhoto }}</span><span @click="loadMorePhoto()">Покажи още</span>
-                </p>
-            </div>
-
+<!--The following <div> is shown only if "photoMethod" parameter is set to true, "pModal" class is added for styling. -->
             <div v-if="photoMethod" :class="'pModal'">
+<!--The following <div> center it's content -->
                 <div class="text-center imgContainer row align-items-center">
                     <div class="col">
-                        <img :class="'img-fluid mt-5 imgContainer'" :src="'/storage/galleries/' + thePhoto.photos" :alt="thePhoto.photos">
+<!--The following <img> is responsive the binded src attribute to the image folder and file with title from "thePhoto" object "photos" field.-->
+                        <img class="img-fluid mt-5" :src="'/storage/images/' + thePhoto.photos" :alt="thePhoto.photos">
                     </div>
                 </div>
-                    <div class="downloadButton">
+                    <div class="downloadButton"><!-- Class "downloadButton" added for styling. -->
+<!--The following <a> has binded href attribute to the image folder, file with title from "thePhoto" object "photos" field and enable download. -->
                         <a class="dwhite" :href="'/storage/galleries/' + thePhoto.photos" download>Изтегли <i class="fas fa-file-download"></i></a>
                     </div>
-                    <div class="closeButton">
-                        <a @click.prevent="singlePhotoModal()"><i class="fas fa-times dwhite fa-3x"></i></a>
+                    <div class="closeButton"><!-- Class "closeButton" added for styling.  -->
+<!--The following <a> has click event with .prevent called to prevent the default behavior of the <a> wich calls "singlePhotoModal" method. -->
+                        <a href="" @click.prevent="singlePhotoModal()"><i class="fas fa-times dwhite fa-3x"></i></a>
                     </div>
-                    <div class="rightButton">
-                        <a @click.prevent="nextPhoto()"><i class="fas fa-chevron-right dwhite fa-3x"></i></a>
+                    <div class="rightButton"><!-- Class "rightButton" added for styling.  -->
+<!--The following <a> has click event with .prevent called to prevent the default behavior of the <a> wich calls "nextPhoto" method. -->
+                        <a href="" @click.prevent="nextPhoto()"><i class="fas fa-chevron-right dwhite fa-3x"></i></a>
                     </div>
-                    <div class="leftButton">
-                        <a @click.prevent="prevPhoto()"><i class="fas fa-chevron-left dwhite fa-3x"></i></a>
+                    <div class="leftButton"><!-- Class "leftButton" added for styling.  -->
+<!--The following <a> has click event with .prevent called to prevent the default behavior of the <a> wich calls "prevPhoto" method. -->
+                        <a href="" @click.prevent="prevPhoto()"><i class="fas fa-chevron-left dwhite fa-3x"></i></a>
                     </div>
-
             </div>
-
-
+<!--The following <div> is created based on "bootstrap" and "vForm" packages -->
             <div class="modal fade" id="photosModal" tabindex="-1" role="dialog" aria-labelledby="photosModalLabel" aria-hidden="true">
                 <div class="modal-dialog modal-xl" role="document">
                     <div v-if="adminMode" class="modal-content">
@@ -51,7 +73,8 @@
                             <span aria-hidden="true">&times;</span>
                             </button>
                         </div>
-                        <form @submit.prevent="createPhoto('this')" id="photosForm" enctype="multipart/form-data">
+<!-- The following <form> when submitted calls "createPhoto" method.-->
+                        <form @submit.prevent="createPhoto()" id="photosForm" enctype="multipart/form-data">
                             <div class="modal-body">
                                 <div class="form-group">
                                     <label>Заглавие:</label>
@@ -65,10 +88,11 @@
                                 </div>
                             </div>
                             <div class="modal-footer">
-                                <button type="submit" v-show="!editMode" class="btn btn-success">Добави</button>
+                                <button type="submit" class="btn btn-success">Добави</button>
                                 <button type="button" class="btn btn-secondary" data-dismiss="modal">Затвори</button>
                             </div>
                         </form>
+
                     </div>
                 </div>
             </div>
@@ -80,38 +104,34 @@
     export default {
         data() {
             return {
-                adminMode : false,
-                editMode : false,
-                photoMethod : false,
-                gallery: null,
-                displayedGallery: null,
-                photoModal : null,
-                photoIndex : null,
-                thePhoto : null,
-                allPhotos : null,
-                loadedPhoto : 16,
-                form : new Form ({
-                    title : null,
-                    photos : null,
+                adminMode : false, // swich to show administrator content
+                photoMethod : false, // swich to show a single photo container
+                gallery: null, // holds the records from the DB
+                displayedGallery: null, //holds the records to display
+                photoIndex : null, //holds the index of a single record
+                thePhoto : null, //holds the signle record
+                allPhotos : null, // holds the number of all the photos get from the DB
+                loadedPhoto : 16, // the numer of photos to display
+                form : new Form ({ // instantiate a new Form object
+                    title : null, // title of/for the DB record
+                    photos : null, // image of/for the DB record
                 }),
             }
         },
         methods: {
-            showModal(){
-                this.editMode = false
+            showModal(){ // show the form modal
                 $("#photosModal").modal("show");
             },
-            changeAdminMode() {
+            changeAdminMode() { //check is the user administrator
                 if(this.$gate.isAdmin()) {
                     this.adminMode = true
                 } else {
                     this.adminMode = false
                 }
             },
-            getPhotos() {
+            getPhotos() { //request to the backend to get the records
                 axios.get("/api/gallery/"+ this.$route.params.id)
                 .then(({ data }) => {
-                    console.log(data)
                     this.gallery = data
                     this.allPhotos = data.length
                     this.displayGallery()
@@ -120,7 +140,7 @@
 
                 });
             },
-            createPhoto(){
+            createPhoto(){ //request to the backend to create a new record
                 if(this.$gate.isAdmin()) {
                     this.form.post('/api/gallery/' + this.$route.params.id)
                     .then(() => {
@@ -134,7 +154,7 @@
                     })
                 }
             },
-            deletePhoto (id) {
+            deletePhoto (id) { // request to the backend to delete a record
                 if(this.$gate.isAdmin()) {
                     axios.delete('/api/gallery/' + id)
                     .then(() => {
@@ -145,7 +165,7 @@
                     })
                 }
             },
-            onFileSelect(event) {
+            onFileSelect(event) { //instantiate new FileReader object for the selected file
                 let file = event.target.files[0];
                 this.form.photos = event.target.files[0];
                 let reader = new FileReader();
@@ -154,10 +174,10 @@
                 }
                 reader.readAsDataURL(file)
             },
-            displayGallery() {
+            displayGallery() {// slice the gallery array to display the desired results
                 this.displayedGallery = this.gallery.slice(0 , this.loadedPhoto)
             },
-            singlePhotoModal(index) {
+            singlePhotoModal(index) { // set the parameters to display a single record
                 if(!this.photoMethod) {
                     this.photoIndex = index
                     this.photoMethod = true
@@ -168,7 +188,7 @@
                     this.thePhoto = null
                 }
             },
-            nextPhoto() {
+            nextPhoto() { // set the parameters to display next single record
                 if(this.photoIndex < this.gallery.length - 1) {
                     this.photoIndex += 1
                     this.thePhoto = this.gallery[this.photoIndex]
@@ -177,7 +197,7 @@
                     this.thePhoto = this.gallery[this.photoIndex]
                 }
             },
-            prevPhoto() {
+            prevPhoto() { // set the parameters to display previous single record
                 if(this.photoIndex >= 1) {
                     this.photoIndex -= 1
                     this.thePhoto = this.gallery[this.photoIndex]
@@ -186,13 +206,12 @@
                     this.thePhoto = this.gallery[this.photoIndex]
                 }
             },
-            loadMorePhoto() {
+            loadMorePhoto() { // set the number of recors to display
                 this.loadedPhoto +=16
                 this.displayGallery()
             }
         },
         mounted() {
-            console.log(this.$route.params)
             this.getPhotos()
             this.changeAdminMode()
         }
